requirements.txt
# Core dependencies
google-adk>=0.2.0
openai>=1.0.0
litellm>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.20.0
requests>=2.28.0
pydantic>=2.0.0

# Utility dependencies
datetime>=4.3
pytz>=2022.1
typing-extensions>=4.0.0
loguru>=0.6.0
